BuildInfoOption.ConstantValue
This is a rework of https://github.com/sbt/sbt-buildinfo/pull/140

Looking closely it doesn't look like `ScalaFinalCaseObjectRenderer` added in #140 is actually going to be useful since it just added `final` in front of all vals.

This provides fix for #127 by generating [constant value definition](https://www.scala-lang.org/files/archive/spec/2.12/04-basic-declarations-and-definitions.html#value-declarations-and-definitions) for `Int`, `Long`, `Double`, `Boolean`, `Symbol`, and `String`.

This allows the values to be used in annotation as the scripted test show. Note due to constant folding using this is not recommended for libraries.